+++ src/SCEMonteCarlo.jl
@@ -71,7 +71,7 @@ export ReducedCell, reduce_cell
 
 public resolve_kt
 public ScaledTerm, SpinConfig, site_index, site_atom
-public site_coeffs!, delta_energy, site_gradient
+public site_coeffs!, delta_energy, site_gradient, energy_gradient, energy_gradient!
 public LogBinner, BinStore, jackknife, std_error, tau_int, bin_means
 public ChainState, SweepScratch, metropolis_sweep!, overrelaxation_sweep!
 public to_matrix, from_matrix